feat(图表): 符号地图大小配置项优化
