+++ core/core-frontend/src/locales/en.ts
@@ -1867,6 +1867,13 @@ Scatter chart (bubble) chart: {a} (series name), {b} (data name), {c} (value arr
     export_raw_details: 'Raw Details',
     field_is_empty_export_error: 'No fields available, unable to export',
     chart_symbolic_map: 'Symbolic map',
+    symbolic: 'Symbolic',
+    symbolic_shape: 'Symbolic Shape',
+    symbolic_upload_hint: 'Supports SVG, JPG, JPEG, PNG files within 1MB',
+    symbolic_range: 'Range',
+    symbolic_error_icon: 'Please select the correct icon file!',
+    symbolic_error_size: 'The file size cannot exceed 1MB!',
+    symbolic_error_range: 'The second range value must be greater than the first range value',
     chart_stock_line: 'K line',
     liquid_condition_style_tips: `Condition style settings, determine water wave chart interval colors, leave blank to disable thresholds, range (0-100), incremental levels<br>Example: input 30,70; this means: divided into 3 segments, namely [0,30], [30,70], [70,100]`,
     conversion_rate: 'Conversion rate',
