+++ core/core-frontend/src/locales/tw.ts
@@ -1827,6 +1827,13 @@ export default {
     export_raw_details: '導出原始明細',
     field_is_empty_export_error: '目前無欄位，無法匯出',
     chart_symbolic_map: '符號地圖',
+    symbolic: '符號',
+    symbolic_shape: '符號形狀',
+    symbolic_upload_hint: '支持 1MB 以內的 SVG, JPG, JPEG, PNG 文件',
+    symbolic_range: '區間',
+    symbolic_error_icon: '請選擇正確的圖標文件!',
+    symbolic_error_size: '文件大小不能超過 1MB!',
+    symbolic_error_range: '第二個區間值必須大于第一個區間值',
     chart_stock_line: 'K 線圖',
     liquid_condition_style_tips: `條件樣式設定，決定水波圖區間顏色，為空則不啟用閾值，範圍(0-100)，逐級遞增<br/>例如：輸入 30,70；表示：分為3段，分別為[0,30],[30,70],[70,100]`,
     conversion_rate: '轉換率',
