+++ core/core-frontend/src/locales/zh-CN.ts
@@ -1829,6 +1829,13 @@ export default {
     export_raw_details: '导出原始明细',
     field_is_empty_export_error: '当前无字段，无法导出',
     chart_symbolic_map: '符号地图',
+    symbolic: '符号',
+    symbolic_shape: '符号形状',
+    symbolic_upload_hint: '支持 1MB 以内的 SVG, JPG, JPEG, PNG 文件',
+    symbolic_range: '区间',
+    symbolic_error_icon: '请选择正确的图标文件!',
+    symbolic_error_size: '文件大小不能超过 1MB!',
+    symbolic_error_range: '第二个区间值必须大于第一个区间值',
     chart_stock_line: 'K 线图',
     liquid_condition_style_tips: `条件样式设置，决定水波图颜色，为空则不开启阈值，范围(0-100)，逐级递增<br/>例如：输入 30,70；表示：分为3段，分别为[0,30],[30,70],[70,100]`,
     conversion_rate: '转化率',
