+++ core/core-frontend/src/views/chart/components/editor/editor-style/ChartStyle.vue
@@ -14,6 +14,7 @@ import { storeToRefs } from 'pinia'
 import CollapseSwitchItem from '@/components/collapse-switch-item/src/CollapseSwitchItem.vue'
 import { ElCollapse, ElCollapseItem } from 'element-plus-secondary'
 import BasicStyleSelector from '@/views/chart/components/editor/editor-style/components/BasicStyleSelector.vue'
+import SymbolicStyleSelector from '@/views/chart/components/editor/editor-style/components/SymbolicStyleSelector.vue'
 import DualBasicStyleSelector from '@/views/chart/components/editor/editor-style/components/DualBasicStyleSelector.vue'
 import ComponentPosition from '@/components/visualization/common/ComponentPosition.vue'
 import BackgroundOverallCommon from '@/components/visualization/component-background/BackgroundOverallCommon.vue'
@@ -363,6 +364,22 @@ watch(
               @onStyleAttrChange="onStyleAttrChange"
             ></common-border-setting>
           </collapse-switch-item>
+
+          <el-collapse-item
+            :effect="themes"
+            name="symbolicStyle"
+            :title="t('chart.symbolic')"
+            v-if="showProperties('symbolic-style-selector')"
+          >
+            <SymbolicStyleSelector
+              :property-inner="propertyInnerAll['symbolic-style-selector']"
+              :themes="themes"
+              :chart="chart"
+              @onBasicStyleChange="onBasicStyleChange"
+              @onMiscChange="onMiscChange"
+            />
+          </el-collapse-item>
+
           <el-collapse-item
             :effect="themes"
             name="events"
