+++ core/core-frontend/src/views/chart/components/editor/editor-style/components/BasicStyleSelector.vue
@@ -253,89 +253,6 @@ const heatMapTypeOptions = [
   { name: t('chart.heatmap3D'), value: 'heatmap3D' }
 ]
 
-const mapSymbolOptions = [
-  { name: t('chart.line_symbol_circle'), value: 'circle' },
-  { name: t('chart.line_symbol_rect'), value: 'square' },
-  { name: t('chart.line_symbol_triangle'), value: 'triangle' },
-  { name: t('chart.map_symbol_pentagon'), value: 'pentagon' },
-  { name: t('chart.map_symbol_hexagon'), value: 'hexagon' },
-  { name: t('chart.map_symbol_octagon'), value: 'octogon' },
-  { name: t('chart.line_symbol_diamond'), value: 'rhombus' },
-  { name: t('commons.custom'), value: 'custom' }
-]
-const iconUpload = ref()
-const acceptedFileType = ['image/svg+xml', 'image/jpeg', 'image/png']
-const onIconChange: UploadProps['onChange'] = async uploadFile => {
-  const rawFile = uploadFile.raw
-  let validIcon = true
-  if (!acceptedFileType.includes(rawFile.type)) {
-    ElMessage.error('请选择正确的图标文件！')
-    validIcon = false
-  }
-  if (rawFile.size / 1024 / 1024 > 1) {
-    ElMessage.error('文件大小不能超过 1MB!')
-    validIcon = false
-  }
-  if (!validIcon) {
-    iconUpload.value?.clearFiles()
-    state.fileList.splice(0)
-    const customIcon = state.basicStyleForm.customIcon
-    if (customIcon) {
-      let file = ''
-      // 图片
-      if (customIcon.startsWith('data')) {
-        file = customIcon
-      } else {
-        // svg
-        file = svgStrToUrl(customIcon)
-      }
-      file && (state.fileList[0] = { url: file })
-    }
-  } else {
-    if (rawFile.type === 'image/svg+xml') {
-      state.basicStyleForm.customIcon = await rawFile.text()
-      changeBasicStyle('customIcon')
-    } else {
-      const fileReader = new FileReader()
-      fileReader.onloadend = () => {
-        state.basicStyleForm.customIcon = fileReader.result as string
-        changeBasicStyle('customIcon')
-      }
-      fileReader.readAsDataURL(rawFile)
-    }
-  }
-}
-
-const changeMapSymbol = () => {
-  const { mapSymbol, customIcon } = state.basicStyleForm
-  if (mapSymbol === 'custom' && customIcon) {
-    let file
-    if (customIcon.startsWith('data')) {
-      file = customIcon
-    } else {
-      file = svgStrToUrl(state.basicStyleForm.customIcon)
-    }
-    file && (state.fileList[0] = { url: file })
-  }
-  changeBasicStyle('mapSymbol')
-}
-
-const customSymbolicMapSizeRange = computed(() => {
-  let { extBubble } = JSON.parse(JSON.stringify(props.chart))
-  return ['symbolic-map'].includes(props.chart.type) && extBubble?.length > 0
-})
-const mapCustomRangeValidate = prop => {
-  const { mapSymbolSizeMax = '0', mapSymbolSizeMin = '1' } = state.basicStyleForm
-  let max = parseInt(mapSymbolSizeMax)
-  let min = parseInt(mapSymbolSizeMin)
-  state.basicStyleForm.mapSymbolSizeMin = Math.max(min, 0)
-  state.basicStyleForm.mapSymbolSizeMax = Math.max(max, 1)
-  if (max < min) {
-    ElMessage.warning('第二个区间值必须大于第一个区间值')
-    return
-  }
-  changeBasicStyle(prop)
-}
 /**
  * 表格是否合并单元格
  */
@@ -571,154 +488,7 @@ onMounted(() => {
         </el-col>
       </el-row>
     </div>
-    <div class="map-flow-style" v-if="showProperty('symbolicMapStyle')">
-      <el-row style="flex: 1">
-        <el-col>
-          <el-form-item class="form-item" :class="'form-item-' + themes">
-            <template #label>
-              <span class="data-area-label">
-                <span style="margin-right: 4px">{{ t('chart.symbolic_map_symbol_shape') }}</span>
-                <el-tooltip class="item" effect="dark" placement="bottom">
-                  <template #content>
-                    <div>{{ t('chart.symbolic_map_symbol_shape_tip') }}</div>
-                  </template>
-                  <el-icon class="hint-icon" :class="{ 'hint-icon--dark': themes === 'dark' }">
-                    <Icon name="icon_info_outlined"><icon_info_outlined class="svg-icon" /></Icon>
-                  </el-icon>
-                </el-tooltip>
-              </span>
-            </template>
-            <el-select
-              :effect="themes"
-              v-model="state.basicStyleForm.mapSymbol"
-              @change="changeMapSymbol()"
-            >
-              <el-option
-                v-for="item in mapSymbolOptions"
-                :key="item.name"
-                :label="item.name"
-                :value="item.value"
-              />
-            </el-select>
-          </el-form-item>
-        </el-col>
-      </el-row>
-      <el-row style="flex: 1" v-if="state.basicStyleForm.mapSymbol === 'custom'">
-        <el-col>
-          <el-form-item class="form-item uploader" :class="'form-item-' + themes">
-            <div class="avatar-uploader-container" :class="`img-area_${themes}`">
-              <el-upload
-                action="#"
-                accept=".svg,.png,.jpeg,.jpg"
-                class="avatar-uploader"
-                list-type="picture-card"
-                ref="iconUpload"
-                :effect="themes"
-                :auto-upload="false"
-                :file-list="state.fileList"
-                :on-change="onIconChange"
-                :limit="1"
-              >
-                <el-icon><Plus /></el-icon>
-              </el-upload>
-            </div>
-          </el-form-item>
-        </el-col>
-      </el-row>
-      <div class="alpha-setting">
-        <label class="alpha-label" :class="{ dark: 'dark' === themes }">
-          {{ t('chart.size') }}
-        </label>
-        <el-row style="flex: 1">
-          <el-col>
-            <el-form-item class="form-item alpha-slider" :class="'form-item-' + themes">
-              <el-slider
-                :effect="themes"
-                :min="1"
-                :max="40"
-                v-model="state.basicStyleForm.mapSymbolSize"
-                @change="changeBasicStyle('mapSymbolSize')"
-                :disabled="customSymbolicMapSizeRange"
-              />
-            </el-form-item>
-          </el-col>
-        </el-row>
-      </div>
-      <div class="alpha-setting">
-        <label class="alpha-label" :class="{ dark: 'dark' === themes }">
-          {{ t('chart.size_range') }}
-        </label>
-        <el-row style="flex: 1">
-          <el-col :span="11">
-            <el-form-item class="form-item alpha-slider" :class="'form-item-' + themes">
-              <el-input
-                type="number"
-                :effect="themes"
-                v-model="state.basicStyleForm.mapSymbolSizeMin"
-                class="basic-input-number"
-                :controls="false"
-                @blur="mapCustomRangeValidate('mapSymbolSizeMin')"
-                :disabled="!customSymbolicMapSizeRange"
-              >
-              </el-input>
-            </el-form-item>
-          </el-col>
-          <el-col :span="1.2">
-            <span>-</span>
-          </el-col>
-          <el-col :span="11">
-            <el-form-item class="form-item alpha-slider" :class="'form-item-' + themes">
-              <el-input
-                type="number"
-                :effect="themes"
-                v-model="state.basicStyleForm.mapSymbolSizeMax"
-                class="basic-input-number"
-                :controls="false"
-                @blur="mapCustomRangeValidate('mapSymbolSizeMax')"
-                :disabled="!customSymbolicMapSizeRange"
-              >
-              </el-input>
-            </el-form-item>
-          </el-col>
-        </el-row>
-      </div>
-      <div v-if="state.basicStyleForm.mapSymbol !== 'custom'" class="alpha-setting">
-        <label class="alpha-label" :class="{ dark: 'dark' === themes }">
-          {{ t('chart.not_alpha') }}
-        </label>
-        <el-row style="flex: 1">
-          <el-col>
-            <el-form-item class="form-item alpha-slider" :class="'form-item-' + themes">
-              <el-slider
-                :effect="themes"
-                :min="1"
-                :max="10"
-                v-model="state.basicStyleForm.mapSymbolOpacity"
-                @change="changeBasicStyle('mapSymbolOpacity')"
-              />
-            </el-form-item>
-          </el-col>
-        </el-row>
-      </div>
-      <div v-if="state.basicStyleForm.mapSymbol !== 'custom'" class="alpha-setting">
-        <label class="alpha-label" :class="{ dark: 'dark' === themes }">
-          {{ t('visualization.borderWidth') }}
-        </label>
-        <el-row style="flex: 1">
-          <el-col>
-            <el-form-item class="form-item alpha-slider" :class="'form-item-' + themes">
-              <el-slider
-                :effect="themes"
-                :min="1"
-                :max="5"
-                v-model="state.basicStyleForm.mapSymbolStrokeWidth"
-                @change="changeBasicStyle('mapSymbolStrokeWidth')"
-              />
-            </el-form-item>
-          </el-col>
-        </el-row>
-      </div>
-    </div>
+
     <!--flow map end-->
     <!--map start-->
     <el-row :gutter="8">
